♻️ BHEAQT release: keep evaluator module below hotspot limit
