+++ .agentplane/tasks/202608131730-BHEAQT/pr/github-body.md
@@ -90,7 +90,8 @@ Publish the 0.7.6 patch only after EZZZYH is merged and closed. Freeze the exact
  ...-cli.critical.agent-efficiency-baseline.test.ts |  57 +++-
  .../src/cli/task-create-planner-intent.testkit.ts  |   2 +
  .../evaluator/evaluator-qualification-review.ts    |   3 +-
- .../commands/evaluator/evaluator-review-usecase.ts |  32 +-
+ .../commands/evaluator/evaluator-review-shared.ts  |  12 +
+ .../commands/evaluator/evaluator-review-usecase.ts |  44 ++-
  packages/agentplane/src/commands/pr/open.ts        |  17 +-
  packages/core/package.json                         |   2 +-
  packages/recipes/package.json                      |   2 +-
@@ -101,7 +102,7 @@ Publish the 0.7.6 patch only after EZZZYH is merged and closed. Freeze the exact
  .../check-compatibility-contract-baseline.mjs      |  59 ++--
  website/static/img/social/docs/releases/v0.7.6.png | Bin 0 -> 53331 bytes
  website/static/img/social/manifest.json            |   8 +
- 74 files changed, 684 insertions(+), 1394 deletions(-)
+ 75 files changed, 698 insertions(+), 1404 deletions(-)
 ```
 
 </details>
