+++ .agentplane/tasks/202608131730-BHEAQT/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202608131730-BHEAQT/qualify-and-publish-agentplane-0-7-6",
   "created_at": "2026-08-13T17:32:00.151Z",
-  "diffstat_sha256": "sha256:b5ab8c75c9bf49542ab4f9d6374dcfd7a4703546a8343c2aa86720455b39ad7b",
+  "diffstat_sha256": "sha256:a2977a14496f975431ae6c700d33d29573339acf54f841e23c95ed56ecf5e7d0",
   "last_verified_at": null,
   "schema_version": 1,
   "task_id": "202608131730-BHEAQT",
